Change variable declaration from unsigned
These patches change the instance where variables or casts are
done with 'unsigned <foo>' to 'unsigned int <foo>' to maintain
consistency with the bulk of code in libpcap-1.3.0.  I know that
in standard C, 'unsigned <foo>' and 'unsigned int <foo>' is the
same, but it should make sense to be consistent throughout the code.